Cache Pyhton objects
Cache the Pyhton objects that wrap Blender objects
to avoid creating multiple wrappers for the same
Blender object. This is necessary as we are
starting to need statefuc functionality in them
(e.g. due to the on-the-fly stitching optimization
and callbacks needed for it.)

The current implementation is bit clunky but does the job for now.